pyproject.toml
[project]
name = "fastapi-tdd"
version = "0.1.0"
description = "Test-Driven Development with FastAPI"
authors = [
    {name = "Carol",email = "awdrg1210@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "ormar[postgresql] (>=0.20.2,<0.21.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
]


[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
ruff = "^0.9.6"
pre-commit = "^4.1.0"
tox = "^4.24.1"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-dotenv = "^0.5.2"
sphinx = "^8.2.3"
sphinx-rtd-theme = "^3.0.2"
sphinx-autodoc-typehints = "^3.2.0"
sphinxcontrib-napoleon = "^0.7" # Ensure this is present and recent
autodoc-pydantic = "^2.2.0"

[tool.ruff]
line-length = 120
extend-exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
